requirements.txt
mysql-connector-python>=8.0.0
pymysql>=1.1.0
pyyaml>=6.0.0

# Web API dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Auth & Security
cryptography>=41.0.0
pyjwt>=2.8.0
